+++ app/Http/Controllers/UserManagement/API/RoleController.php
@@ -4,13 +4,17 @@ namespace App\Http\Controllers\UserManagement\API;
 
 use App\Http\Controllers\Controller;
 use App\Http\Controllers\ValidatesResourceIndex;
-use App\Http\Requests\UserManagement\StoreRole;
+use App\Http\Requests\UserManagement\StoreUpdateRole;
 use App\Http\Resources\Role as RoleResource;
 use App\Http\Resources\RoleCollection;
 use App\Http\Resources\UserCollection;
 use App\Models\Role;
 use App\Models\User;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
+use Illuminate\Http\Resources\Json\JsonResource;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Log;
 
 class RoleController extends Controller
 {
@@ -21,12 +25,7 @@ class RoleController extends Controller
         $this->authorizeResource(Role::class);
     }
 
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
+    public function index(): JsonResource
     {
         $this->validateFilter();
         $this->validateSorting([
@@ -41,18 +40,18 @@ class RoleController extends Controller
             ->get());
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(StoreRole $request)
+    public function store(StoreUpdateRole $request): JsonResponse
     {
         $role = new Role();
-        $role->fill($request->all());
+        $role->fill($request->validated());
 
         $role->save();
+
+        Log::info('User role has been created.', [
+            'role_id' => $role->id,
+            'role_name' => $role->name,
+            'client_ip' => $request->ip(),
+        ]);
 
         return response()
             ->json([
@@ -61,29 +60,22 @@ class RoleController extends Controller
             ->header('Location', route('api.roles.show', $role));
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\Role  $role
-     * @return \Illuminate\Http\Response
-     */
-    public function show(Role $role)
+    public function show(Role $role): JsonResource
     {
         return new RoleResource($role);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Role  $role
-     * @return \Illuminate\Http\Response
-     */
-    public function update(StoreRole $request, Role $role)
+    public function update(StoreUpdateRole $request, Role $role): JsonResponse
     {
-        $role->fill($request->all());
+        $role->fill($request->validated());
 
         $role->save();
+
+        Log::info('User role has been updated.', [
+            'role_id' => $role->id,
+            'role_name' => $role->name,
+            'client_ip' => $request->ip(),
+        ]);
 
         return response()
             ->json([
@@ -91,15 +83,15 @@ class RoleController extends Controller
             ]);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Role  $role
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy(Role $role)
+    public function destroy(Request $request, Role $role): JsonResponse
     {
         $role->delete();
+
+        Log::info('User role has been deleted.', [
+            'role_id' => $role->id,
+            'role_name' => $role->name,
+            'client_ip' => $request->ip(),
+        ]);
 
         return response()
             ->json([
@@ -107,12 +99,7 @@ class RoleController extends Controller
             ]);
     }
 
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function users(Role $role)
+    public function users(Role $role): JsonResource
     {
         $this->authorize('view', $role);
         $this->authorize('viewAny', User::class);
@@ -122,12 +109,7 @@ class RoleController extends Controller
             ->get());
     }
 
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function administrators(Role $role)
+    public function administrators(Role $role): JsonResource
     {
         $this->authorize('view', $role);
         $this->authorize('viewAny', User::class);