+++ app/Http/Controllers/UserManagement/API/UserRoleRelationshipController.php
@@ -5,35 +5,26 @@ namespace App\Http\Controllers\UserManagement\API;
 use App\Http\Controllers\Controller;
 use App\Models\Role;
 use App\Models\User;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 
 class UserRoleRelationshipController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index(User $user)
+    public function index(User $user): JsonResponse
     {
         $this->authorize('view', $user);
         $this->authorize('viewAny', Role::class);
 
-        return [
-            'data' => [
-                'id' => $user->roles->pluck('id'),
-            ],
-        ];
+        return response()
+            ->json([
+                'data' => [
+                    'id' => $user->roles->pluck('id'),
+                ],
+            ]);
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(User $user, Request $request)
+    public function store(User $user, Request $request): JsonResponse
     {
         $this->authorize('update', $user);
 
@@ -56,13 +47,7 @@ class UserRoleRelationshipController extends Controller
             ], Response::HTTP_CREATED);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function update(User $user, Request $request)
+    public function update(User $user, Request $request): JsonResponse
     {
         $this->authorize('update', $user);
 
@@ -84,13 +69,7 @@ class UserRoleRelationshipController extends Controller
             ]);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Role  $role
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy(User $user, Request $request)
+    public function destroy(User $user, Request $request): JsonResponse
     {
         $this->authorize('update', $user);