+++ app/Http/Controllers/UserManagement/UserController.php
@@ -3,15 +3,17 @@
 namespace App\Http\Controllers\UserManagement;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\UserManagement\StoreUser;
-use App\Http\Requests\UserManagement\UpdateUser;
+use App\Http\Requests\UserManagement\StoreUpdateUser;
 use App\Models\Role;
 use App\Models\User;
 use App\Util\AutoColorInitialAvatar;
 use App\View\Components\UserAvatar;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Log;
+use Illuminate\View\View;
 use Storage;
 
 class UserController extends Controller
@@ -21,41 +23,23 @@ class UserController extends Controller
         $this->authorizeResource(User::class);
     }
 
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index(Request $request)
+    public function index(Request $request): View
     {
         return view('vue-app');
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
+    public function create(): View
     {
         return view('user_management.users.create', [
             'roles' => Role::orderBy('name')->get(),
         ]);
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \App\Http\Requests\UserManagement\StoreUser  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(StoreUser $request)
+    public function store(StoreUpdateUser $request): RedirectResponse
     {
         $user = new User();
-        $user->name = $request->name;
-        $user->email = $request->email;
+        $user->fill($request->validated());
         $user->password = Hash::make($request->password);
-        $user->is_super_admin = ! empty($request->is_super_admin);
         $user->save();
         $user->roles()->sync($request->roles);
 
@@ -63,7 +47,7 @@ class UserController extends Controller
             'user_id' => $user->id,
             'user_name' => $user->name,
             'email' => $user->email,
-            'client_ip' => request()->ip(),
+            'client_ip' => $request->ip(),
         ]);
 
         return redirect()
@@ -71,24 +55,12 @@ class UserController extends Controller
             ->with('success', __('User has been added.'));
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  User  $user
-     * @return \Illuminate\Http\Response
-     */
-    public function show(User $user)
+    public function show(User $user): View
     {
         return view('vue-app');
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  User  $user
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(User $user)
+    public function edit(User $user): View
     {
         return view('user_management.users.edit', [
             'user' => $user,
@@ -96,20 +68,10 @@ class UserController extends Controller
         ]);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \App\Http\Requests\UserManagement\UpdateUser  $request
-     * @param  User  $user
-     * @return \Illuminate\Http\Response
-     */
-    public function update(UpdateUser $request, User $user)
+    public function update(StoreUpdateUser $request, User $user): RedirectResponse
     {
-        $user->name = $request->name;
-        if (empty($user->provider_name)) {
-            $user->email = $request->email;
-        }
-        if ($request->password !== null) {
+        $user->fill($request->validated());
+        if ($request->filled('password')) {
             $user->password = Hash::make($request->password);
         }
         $user->is_super_admin = ! empty($request->is_super_admin) || User::count() == 1;
@@ -121,7 +83,7 @@ class UserController extends Controller
             'user_id' => $user->id,
             'user_name' => $user->name,
             'email' => $user->email,
-            'client_ip' => request()->ip(),
+            'client_ip' => $request->ip(),
         ]);
 
         return redirect()
@@ -129,13 +91,7 @@ class UserController extends Controller
             ->with('success', __('User has been updated.'));
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  User  $user
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy(User $user)
+    public function destroy(Request $request, User $user): RedirectResponse
     {
         if ($user->avatar !== null && Storage::exists($user->avatar)) {
             Storage::delete($user->avatar);
@@ -147,7 +103,7 @@ class UserController extends Controller
             'user_id' => $user->id,
             'user_name' => $user->name,
             'email' => $user->email,
-            'client_ip' => request()->ip(),
+            'client_ip' => $request->ip(),
         ]);
 
         return redirect()
@@ -155,10 +111,7 @@ class UserController extends Controller
             ->with('success', __('User has been deleted.'));
     }
 
-    /**
-     * Lists all permissions
-     */
-    public function permissions()
+    public function permissions(): View
     {
         $this->authorize('viewAny', User::class);
 
@@ -167,13 +120,7 @@ class UserController extends Controller
         ]);
     }
 
-    /**
-     * Display the avatar of the user.
-     *
-     * @param  \App\Models\User  $user
-     * @return \Illuminate\Http\Response
-     */
-    public function avatar(User $user, Request $request)
+    public function avatar(User $user, Request $request): Response
     {
         $request->validate([
             'size' => [