+++ app/Http/Requests/UserManagement/UpdateMembers.php
@@ -9,11 +9,21 @@ use Illuminate\Validation\Rule;
 class UpdateMembers extends FormRequest
 {
     /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    /**
      * Get the validation rules that apply to the request.
      *
      * @return array
      */
-    public function rules()
+    public function rules(): array
     {
         return [
             'users' => [
@@ -21,15 +31,5 @@ class UpdateMembers extends FormRequest
                 Rule::in(User::select('id')->get()->pluck('id')),
             ],
         ];
-    }
-
-    /**
-     * Determine if the user is authorized to make this request.
-     *
-     * @return bool
-     */
-    public function authorize()
-    {
-        return true;
     }
 }