+++ resources/js/components/userprofile/AccountDeleteDialog.vue
@@ -2,11 +2,16 @@
     <b-card
         class="shadow-sm mb-4"
         :header="$t('Account Removal')"
-        body-class="pb-1"
         footer-class="text-right"
+        :no-body="!canDelete"
     >
-        <p>{{ $t('If you no longer plan to use this service, you can remove your account and delete all associated data.') }}</p>
-        <template #footer>
+        <b-card-text v-if="canDelete">
+            {{ $t('If you no longer plan to use this service, you can remove your account and delete all associated data.') }}
+        </b-card-text>
+        <b-alert v-else show variant="info" class="m-0">
+            {{ $t('Account cannot be deleted as it is the only remaining account with super-admin privileges.') }}
+        </b-alert>
+        <template v-if="canDelete" #footer>
             <b-button
                 type="button"
                 variant="danger"
@@ -24,6 +29,9 @@
 import userprofileApi from "@/api/userprofile"
 import { showSnackbar } from '@/utils'
 export default {
+    props: {
+        canDelete: Boolean,
+    },
     data() {
         return {
             isBusy: false,