+++ tests/Feature/RoleUserApiTest.php
@@ -26,6 +26,7 @@ class RoleUserApiTest extends TestCase
     {
         $role = Role::factory()->create();
 
+        /** @var User $authUser */
         $authUser = User::factory()->make();
 
         $response = $this->actingAs($authUser)
@@ -41,6 +42,7 @@ class RoleUserApiTest extends TestCase
 
         $authUser = $this->makeUserWithPermission('app.usermgmt.view');
 
+        /** @var User $authUser */
         $response = $this->actingAs($authUser)
             ->getJson('api/roles/'. 1234 .'/users', []);
 
@@ -55,6 +57,7 @@ class RoleUserApiTest extends TestCase
     {
         $role = Role::factory()->create();
 
+        /** @var User $authUser */
         $authUser = $this->makeUserWithPermission('app.usermgmt.view');
 
         $response = $this->actingAs($authUser)
@@ -80,6 +83,7 @@ class RoleUserApiTest extends TestCase
 
         $authUser = $this->makeUserWithPermission('app.usermgmt.view');
 
+        /** @var User $authUser */
         $response = $this->actingAs($authUser)
             ->getJson('api/roles/'.$role->id.'/users', []);
 
@@ -94,7 +98,10 @@ class RoleUserApiTest extends TestCase
                         'locale' => $user1->locale,
                         'is_super_admin' => $user1->is_super_admin,
                         'is_2fa_enabled' => false,
+                        'is_current_user' => false,
                         'avatar_url' => $user1->avatarUrl(),
+                        'can_update' => false,
+                        'can_delete' => false,
                         'provider_name' => $user1->provider_name,
                         'created_at' => $user1->created_at->toJSON(),
                         'updated_at' => $user1->updated_at->toJSON(),
@@ -105,6 +112,12 @@ class RoleUserApiTest extends TestCase
                             'edit' => route('users.edit', $user1),
                         ],
                         'relationships' => [
+                            'administeredRoles' => [
+                                'links' => [
+                                    'related' => route('api.users.roles.index', $user1),
+                                    'self' => route('api.users.relationships.roles.index', $user1),
+                                ],
+                            ],
                             'roles' => [
                                 'links' => [
                                     'related' => route('api.users.roles.index', $user1),
@@ -120,7 +133,10 @@ class RoleUserApiTest extends TestCase
                         'locale' => $user2->locale,
                         'is_super_admin' => $user2->is_super_admin,
                         'is_2fa_enabled' => false,
+                        'is_current_user' => false,
                         'avatar_url' => $user2->avatarUrl(),
+                        'can_update' => false,
+                        'can_delete' => false,
                         'provider_name' => $user2->provider_name,
                         'created_at' => $user2->created_at->toJSON(),
                         'updated_at' => $user2->updated_at->toJSON(),
@@ -131,6 +147,12 @@ class RoleUserApiTest extends TestCase
                             'edit' => route('users.edit', $user2),
                         ],
                         'relationships' => [
+                            'administeredRoles' => [
+                                'links' => [
+                                    'related' => route('api.users.roles.index', $user2),
+                                    'self' => route('api.users.relationships.roles.index', $user2),
+                                ],
+                            ],
                             'roles' => [
                                 'links' => [
                                     'related' => route('api.users.roles.index', $user2),